Run heavy orbital/collision search in a web worker
The 3D orbit-proximity + contact-window collision search
(OrbitalRelationsCore.detectCollisionStatus and friends) ran synchronously
on the main thread in SystemBodyComponent.ngOnChanges, janking the UI on
collision-heavy systems. Move it off-thread via a shared Comlink web worker,
establishing a reusable offload pattern for the heavier calculations to come.

- Extract the engine into framework-free orbital-relations.core.ts;
  OrbitalRelationsService now just `extends` it for DI and re-exports its types.
- Add OrbitalWorkerService (lazy shared worker, Comlink, inline fallback) plus
  collision.worker.ts / collision-worker-api.ts, and a serialize/rehydrate
  boundary (collision-request.ts) so the cyclic SystemBody tree isn't posted whole.
- SystemBodyComponent: collisionStatus/collisionPending are signals with a
  per-request generation token (drops stale/post-destroy responses) and a
  150ms-gated pending skeleton; showCollisionDialog awaits the worker.
- Tests for the boundary, facade (inline + worker paths) and the Comlink wire;
  coverage held >=80%. Document the pattern in AGENTS.md.
