+++ package.json
@@ -28,6 +28,7 @@
     "@fortawesome/angular-fontawesome": "^5.0.0",
     "@fortawesome/fontawesome-svg-core": "^6.7.2",
     "@fortawesome/free-solid-svg-icons": "^6.7.2",
+    "comlink": "^4.4.2",
     "rxjs": "~7.8.2",
     "tslib": "^2.8.1"
   },
