+++ src/app/data/orbital-relations.service.ts
@@ -1,1060 +1,21 @@
 import { Injectable } from '@angular/core';
-import { CanonnBiostatsBody, SystemBody } from '../home/home.component';
-import { BODY_TYPE } from './body-types';
-
-/** The five Lagrange points of a two-body system. */
-export type LagrangePointId = 'L1' | 'L2' | 'L3' | 'L4' | 'L5';
-
-/** Result of Trojan/Lagrange analysis for a body relative to its co-orbital siblings. */
-export interface TrojanStatus {
-  /** Lagrange point the body occupies ('L1'–'L5'), or null when it is not a Trojan. */
-  lagrangePoint: LagrangePointId | null;
-  /** True when this body hosts Trojans at both L4 and L5 (it is the reference body, not a Trojan). */
-  isHost: boolean;
-}
-
-/** A body sitting at one of the slots of a Lagrange configuration. */
-export interface LagrangeOccupant {
-  /** The body's name (raw — callers may map it to a display name). */
-  name: string;
-  /** Its `bodyId`, for DOM ids / de-duplication. */
-  bodyId: number;
-  /** True for the body the configuration was requested for (so the UI can highlight it). */
-  isFocus: boolean;
-}
+import { OrbitalRelationsCore } from './orbital-relations.core';
 
 /**
- * A co-orbital family expressed as a Lagrange diagram: the central primary they orbit,
- * the secondary (a host with companions at L4 & L5, when one exists), and whichever
- * bodies occupy each Lagrange point. Built by {@link OrbitalRelationsService.lagrangeConfiguration}.
- */
-export interface LagrangeConfiguration {
-  /** Name of the shared parent the family orbits (the central primary), or null. */
-  primaryName: string | null;
-  /** The co-orbital secondary (host), or null when none is recorded → drawn as a placeholder. */
-  secondary: LagrangeOccupant | null;
-  /** Bodies at each Lagrange point — usually 0 or 1 per point. */
-  points: Record<LagrangePointId, LagrangeOccupant[]>;
-}
-
-/** A future periapsis/apoapsis passage: when it occurs and how many days away it is. */
-export interface OrbitalEvent {
-  date: Date;
-  days: number;
-}
-
-/** A single (time, centre-to-centre distance) sample for plotting separation against time. */
-export interface SeparationSample {
-  /** Epoch milliseconds of the sample. */
-  tMs: number;
-  /** Centre-to-centre distance (km) between the two bodies at {@link tMs}. */
-  sepKm: number;
-}
-
-/**
- * The contact *window* of a collision: a collision is an interval, not an instant. `start`
- * is when the bodies' separation first drops below the sum of their radii and `end` is when
- * it rises back above; `days` is the number of days from `now` until `start` (it can be
- * slightly negative when contact is already in progress at `now`).
- */
-export interface CollisionWindow {
-  start: Date;
-  end: Date;
-  days: number;
-  /** Centre-to-centre separation (km) at closest approach within the window — the deepest point of contact. */
-  minSeparationKm: number;
-  /**
-   * Name of the sibling body this contact is with. Always set by {@link OrbitalRelationsService.detectCollisionStatus}
-   * (a body can have several crossing partners); optional only so hand-built test fixtures may omit it.
-   */
-  partnerName?: string;
-  /**
-   * Sum of the two bodies' radii (km) for *this* pair — the contact threshold used to gauge overlap.
-   * Distinct per partner, so it travels with the window rather than living once on the status.
-   */
-  combinedRadiiKm?: number;
-}
-
-/** Result of collision-candidate analysis for a body relative to a crossing-orbit sibling. */
-export interface CollisionStatus {
-  /** True when this body shares its parent with a sibling on a crossing, near-coplanar orbit. */
-  isCandidate: boolean;
-  /**
-   * Name of the *primary* partner — the sibling owning the soonest predicted collision (or the
-   * first geometric candidate when no pair has timing data). Null when there is no partner.
-   * A body may collide with several siblings; see {@link upcomingCollisions} for the full set.
-   */
-  partnerName: string | null;
-  /** Synodic period (days) with the primary partner — the interval between successive conjunctions. */
-  synodicPeriodDays: number | null;
-  /** Next contact window (when any pair is within combined radii), or null when timing data is missing. */
-  nextCollision: CollisionWindow | null;
-  /**
-   * Up to 10 upcoming contact windows in chronological order, merged across every crossing
-   * partner of this body; empty when timing data is missing. Each window names its own partner.
-   */
-  upcomingCollisions: CollisionWindow[];
-  /** Sum of the two bodies' radii (km) — the contact threshold; null when not a candidate. */
-  combinedRadiiKm: number | null;
-  /**
-   * Names of additional siblings (beyond the primary partner) that are also on crossing orbits
-   * with any member of this collision group, forming a multi-body collision cluster.
-   * Empty for simple pairs.
-   */
-  simultaneousPartners: string[];
-}
-
-/**
- * A timed multi-body pile-up: an interval in which a reference body is simultaneously in
- * contact with two or more siblings. Detected over a fixed time horizon rather than from the
- * (capped) upcoming-contacts list, so a cluster further out than the listed rows is still found.
- */
-export interface SimultaneousCollision {
-  /** Full names of the sibling bodies the reference body contacts at once (≥ 2). */
-  partnerNames: string[];
-  /** Earliest contact start across the cluster. */
-  start: Date;
-  /** Latest contact end across the cluster. */
-  end: Date;
-  /** Days from `now` until the cluster opens (slightly negative when already in progress). */
-  days: number;
-}
-
-/** Milliseconds per day, used to convert orbital periods (days) to wall-clock time. */
-const MS_PER_DAY = 1000 * 60 * 60 * 24;
-
-/** Kilometres per astronomical unit, to compare body radii against orbital distances. */
-const KM_PER_AU = 149597870.7;
-/** Radians per degree, for the 3D Keplerian position maths. */
-const DEG_TO_RAD = Math.PI / 180;
-/**
- * Number of evenly-spaced mean-anomaly samples per orbit for the coarse proximity scan.
- * The coarse grid only has to resolve the *basins* of the orbit-to-orbit distance surface
- * (which are far wider than the contact width), not the contact arc itself: the narrow
- * close-approach valley of two tightly-nested, tilted orbits is caught separately by the
- * line-of-nodes seeds, and every seed is then zoomed to sub-km precision by
- * {@link OrbitalRelationsService.refineOrbitMinimum}. So we cap the grid at a fixed sample
- * count (0.5° spacing — four times finer than the historical 1° that missed valleys, and
- * the node seeds now cover those valleys regardless) instead of scaling the step down to the
- * contact arc. The old adaptive step bottomed out at 0.05° for any orbit beyond ~0.1 AU,
- * driving the O(N²) coarse double-loop to 7200² ≈ 52M iterations (~200 ms) per sibling pair.
- */
-const ORBIT_COARSE_SAMPLES = 720;
-/** How many of the closest coarse cells to refine from (covers multiple close-approach basins). */
-const ORBIT_REFINE_TOPK = 6;
-/** Half-grid size (per axis) for each zoom pass refining the orbit-to-orbit minimum. */
-const ORBIT_REFINE_GRID = 4;
-/** Number of zoom passes refining the orbit-to-orbit minimum distance. */
-const ORBIT_REFINE_ITERATIONS = 10;
-/** Upper bound on conjunctions examined before giving up on finding a collision date. */
-const MAX_CONJUNCTIONS_SCANNED = 300;
-
-/** Cap on how many upcoming contact windows are surfaced (merged across all crossing partners). */
-const MAX_UPCOMING_CONTACTS = 10;
-
-/** Angular tolerance (degrees) for matching a Lagrange geometry. */
-const ANGLE_TOLERANCE_DEG = 1;
-/** Tolerance (degrees) for L1/L2 alignment of argument-of-periapsis and ascending node. */
-const ALIGNMENT_TOLERANCE_DEG = 5;
-/** Tolerance (degrees) for equal rosette spacing. */
-const ROSETTE_TOLERANCE_DEG = 5;
-
-/** A 3D position in the system's shared frame (kilometres). */
-interface Vec3 { x: number; y: number; z: number; }
-
-/**
- * Detects co-orbital configurations (Trojan/Lagrange points and rosettes) by comparing
- * a body's Keplerian elements with those of its siblings under the same parent. Pure —
- * no Angular/DOM dependency — so it can be unit-tested directly. Extracted from
- * SystemBodyComponent.
+ * Angular-injectable facade over {@link OrbitalRelationsCore}. All the orbital maths lives in
+ * the framework-free core so the collision web worker can import it without pulling any Angular
+ * runtime into the worker bundle; this subclass exists only to provide `@Injectable` DI on the
+ * main thread and to re-export the core's result types under the original module path, so
+ * existing `import { CollisionWindow, LagrangeConfiguration, … } from './orbital-relations.service'`
+ * sites keep resolving unchanged.
+ *
+ * The heavy collision methods (`detectCollisionStatus`, `simultaneousCollisionsWithin`,
+ * `upcomingContactsWithin`, `separationSeries`) remain callable here synchronously — which is how
+ * they stay unit-tested — but at runtime {@link SystemBodyComponent} runs them off the main thread
+ * through {@link OrbitalWorkerService}. The light methods (trojan/rosette/anomaly/Lagrange) stay on
+ * the main thread via this service.
  */
 @Injectable({ providedIn: 'root' })
-export class OrbitalRelationsService {
-  /** Bodies that share this body's parent (excluding itself) and expose an argOfPeriapsis. */
-  private coOrbitalSiblings(
-    body: SystemBody,
-    predicate: (sibling: SystemBody) => boolean,
-  ): SystemBody[] {
-    const parent = body.parent;
-    if (!parent) { return []; }
-    return parent.subBodies.filter(sibling =>
-      sibling !== body &&
-      sibling.bodyData.orbitalPeriod === body.bodyData.orbitalPeriod &&
-      sibling.bodyData.argOfPeriapsis !== undefined &&
-      predicate(sibling),
-    );
-  }
+export class OrbitalRelationsService extends OrbitalRelationsCore {}
 
-  /** Signed angular difference a − b normalised to (−180, 180]. */
-  private signedAngleDiff(a: number, b: number): number {
-    return ((a - b + 540) % 360) - 180;
-  }
-
-  detectTrojanStatus(body: SystemBody): TrojanStatus {
-    const result: TrojanStatus = { lagrangePoint: null, isHost: false };
-    const bd = body.bodyData;
-    if (!body.parent || !bd.orbitalPeriod || !bd.semiMajorAxis || bd.argOfPeriapsis === undefined) {
-      return result;
-    }
-
-    // L1, L2 and L3 are defined relative to a real massive central body. When the shared
-    // parent is a barycentre there is no such third body — its children are the components
-    // of a binary (or families) orbiting a centre of mass — so those points are suppressed
-    // below (the L3 branch and the L1/L2 block both bail on `parentIsBarycentre`). A ±60°
-    // Trojan host, by contrast, *is* the massive co-orbital secondary its companions hang
-    // off, so it remains valid even around a barycentre; we therefore no longer suppress the
-    // whole family here (commit 97d9a50 did, which also hid genuine barycentre-orbiting
-    // Trojan hosts like Hyuqoae GH-V f2-368 AB 2/3/4).
-    const parentIsBarycentre = body.parent.bodyData.type === BODY_TYPE.Barycentre;
-
-    // L3, L4, L5 candidates share the same orbital distance (semi-major axis).
-    //
-    // A body's true along-orbit position is its mean longitude λ = Ω + ω + M
-    // (ascendingNode + argOfPeriapsis + meanAnomaly), so the angular separation of two
-    // co-orbital siblings is Δλ = ΔΩ + Δω + ΔM. We compare argOfPeriapsis (Δω) alone.
-    // That is exact — not merely a heuristic — for Elite's data, because the game holds
-    // Ω and M *identical* across co-orbital siblings and encodes the entire along-orbit
-    // offset in argOfPeriapsis, so ΔΩ = ΔM = 0 and Δλ = Δω. Verified against every real
-    // co-orbital pair in the fixtures — both 180° binaries (Alpha Centauri's "2045 PC2" /
-    // "Lagrange" L3 pair, Merope 1 a/b, …) and genuine ±60° Trojans (Pro Eurl JF-A d88,
-    // Pipe (stem) Sector DL-Y d17, Prooe Bli FQ-R c19-2, Truecho NE-P c22-0, and the
-    // Eorld Byio AA-A h539 host with companions at both L4 and L5): in each, ascendingNode
-    // and meanAnomaly match between siblings (or are absent on both) and only argOfPeriapsis
-    // differs. See the orbital-relations spec's "real game fixtures" block for the pinned values.
-    const sameSMABodies = this.coOrbitalSiblings(body, sibling =>
-      sibling.bodyData.semiMajorAxis === bd.semiMajorAxis,
-    );
-
-    // A body with co-orbital neighbours at both +60° and −60° is the host of the Trojan
-    // pair (the massive reference body); it should not itself be labelled a Trojan.
-    let hasLeadingTrojan = false;
-    let hasTrailingTrojan = false;
-    for (const sibling of sameSMABodies) {
-      const diff = this.signedAngleDiff(sibling.bodyData.argOfPeriapsis!, bd.argOfPeriapsis!);
-      if (Math.abs(diff - 60) < ANGLE_TOLERANCE_DEG) hasLeadingTrojan = true;
-      if (Math.abs(diff + 60) < ANGLE_TOLERANCE_DEG) hasTrailingTrojan = true;
-    }
-    if (hasLeadingTrojan && hasTrailingTrojan) {
-      result.isHost = true;
-      return result;
-    }
-
-    for (const sibling of sameSMABodies) {
-      const argDiff = Math.abs(bd.argOfPeriapsis! - sibling.bodyData.argOfPeriapsis!);
-      const normalizedDiff = Math.min(argDiff, 360 - argDiff);
-
-      if (Math.abs(normalizedDiff - 60) < ANGLE_TOLERANCE_DEG) {
-        const relativePos = this.signedAngleDiff(bd.argOfPeriapsis!, sibling.bodyData.argOfPeriapsis!);
-        result.lagrangePoint = relativePos > 0 ? 'L4' : 'L5';
-        return result;
-      } else if (!parentIsBarycentre && Math.abs(normalizedDiff - 180) < ANGLE_TOLERANCE_DEG) {
-        // 180° opposition → L3, but only around a real central body. A 180° pair orbiting a
-        // barycentre is a binary, not an L3, so it falls through here and is left unbadged.
-        result.lagrangePoint = 'L3';
-        return result;
-      }
-    }
-
-    // L1 and L2 (like L3) are defined relative to a real massive central body, so under a
-    // barycentre none of them apply — only the ±60° Trojan host detected above (which is
-    // itself the massive co-orbital secondary) survives. A binary's components are 180°
-    // opposed, never aligned, so the L1/L2 test below can't misfire on one anyway, but we
-    // stop here to keep the barycentre rule symmetric with the suppressed L3.
-    if (parentIsBarycentre) {
-      return result;
-    }
-
-    // L1, L2 share the orbital period but sit at a different distance, aligned in
-    // argument-of-periapsis and ascending node.
-    const samePeriodBodies = this.coOrbitalSiblings(body, sibling =>
-      sibling.bodyData.semiMajorAxis !== bd.semiMajorAxis &&
-      sibling.bodyData.ascendingNode !== undefined,
-    );
-
-    for (const sibling of samePeriodBodies) {
-      // Use wrapped angular differences so an aligned pair straddling the 0°/360°
-      // seam (e.g. 1° vs 359°) is still recognised as aligned.
-      const argDiff = Math.abs(this.signedAngleDiff(bd.argOfPeriapsis!, sibling.bodyData.argOfPeriapsis!));
-      const nodeDiff = Math.abs(this.signedAngleDiff(bd.ascendingNode || 0, sibling.bodyData.ascendingNode || 0));
-
-      if (argDiff < ALIGNMENT_TOLERANCE_DEG && nodeDiff < ALIGNMENT_TOLERANCE_DEG) {
-        result.lagrangePoint = bd.semiMajorAxis! < sibling.bodyData.semiMajorAxis! ? 'L1' : 'L2';
-        return result;
-      }
-    }
-
-    return result;
-  }
-
-  /**
-   * Resolves the whole co-orbital family `body` belongs to into a Lagrange diagram: the
-   * shared parent (central primary), the secondary host (if any), and which bodies occupy
-   * each Lagrange point. Every family member is classified with {@link detectTrojanStatus},
-   * so the assignment matches the badges shown on each body. `body` itself is flagged as the
-   * focus so the UI can highlight it.
-   *
-   * Returns null when there is nothing meaningful to draw — no parent, no shared orbital
-   * period, or no co-orbital relationships detected anywhere in the family.
-   */
-  lagrangeConfiguration(body: SystemBody): LagrangeConfiguration | null {
-    const parent = body.parent;
-    const bd = body.bodyData;
-    if (!parent || !bd.orbitalPeriod) {
-      return null;
-    }
-
-    // A barycentre can be the centre of a genuine Trojan configuration (its mass hosts a
-    // ±60° co-orbital secondary at L4/L5), so — unlike commit 97d9a50, which bailed out for
-    // every barycentre — we build the diagram and let detectTrojanStatus decide per member.
-    // A plain 180° binary orbiting a barycentre yields no classified members (its L3 is
-    // suppressed there), so the family below has no occupants and we still return null.
-
-    // The co-orbital family: every sibling sharing this body's orbital period (so L1/L2
-    // partners at a different radius are included too), plus `body` itself — it already
-    // lives in `parent.subBodies`. Members without an argOfPeriapsis can't be classified.
-    const family = parent.subBodies.filter(member =>
-      member.bodyData.orbitalPeriod === bd.orbitalPeriod &&
-      member.bodyData.argOfPeriapsis !== undefined,
-    );
-
-    const points: Record<LagrangePointId, LagrangeOccupant[]> = { L1: [], L2: [], L3: [], L4: [], L5: [] };
-    let secondary: LagrangeOccupant | null = null;
-
-    for (const member of family) {
-      const status = this.detectTrojanStatus(member);
-      if (!status.isHost && !status.lagrangePoint) { continue; }
-      const occupant: LagrangeOccupant = {
-        name: member.bodyData.name,
-        bodyId: member.bodyData.bodyId,
-        isFocus: member === body,
-      };
-      if (status.isHost) {
-        secondary = occupant;
-      } else if (status.lagrangePoint) {
-        points[status.lagrangePoint].push(occupant);
-      }
-    }
-
-    // An L3 opposition has no host (neither body has companions at both L4 and L5), so the
-    // pair both read as L3 and the secondary slot would be left empty — stacking the two
-    // opposed bodies on the single L3 marker. But L3 is defined as the point 180° across the
-    // primary *from the secondary*, so one of the opposed pair plays the secondary's role:
-    // promote it into the secondary slot and the diagram draws them on opposite sides of the
-    // orbit (secondary at 0°, the other at L3), exactly the geometry the configuration is.
-    // Promote a non-focused member when possible, so a dialog opened from an L3 badge keeps
-    // the clicked body on the labelled L3 marker.
-    if (secondary === null && points.L3.length > 1) {
-      const focusIndex = points.L3.findIndex(o => o.isFocus);
-      const promoteIndex = focusIndex === 0 ? 1 : 0;
-      [secondary] = points.L3.splice(promoteIndex, 1);
-    }
-
-    const hasOccupant = secondary !== null || Object.values(points).some(slot => slot.length > 0);
-    if (!hasOccupant) {
-      return null;
-    }
-
-    return { primaryName: parent.bodyData.name, secondary, points };
-  }
-
-  /**
-   * Returns a "Rosette (n)" label when this body belongs to a group of ≥3 co-orbital
-   * bodies evenly spaced around the parent, or null otherwise.
-   */
-  detectRosetteStatus(body: SystemBody): string | null {
-    const bd = body.bodyData;
-    if (!body.parent || !bd.orbitalPeriod || !bd.semiMajorAxis || bd.argOfPeriapsis === undefined) {
-      return null;
-    }
-
-    // Include this body itself in the group (it shares its own elements).
-    const rosetteGroup = body.parent.subBodies.filter(sibling =>
-      sibling.bodyData.orbitalPeriod === bd.orbitalPeriod &&
-      sibling.bodyData.semiMajorAxis === bd.semiMajorAxis &&
-      sibling.bodyData.argOfPeriapsis !== undefined,
-    );
-
-    if (rosetteGroup.length < 3) return null;
-
-    const angles = rosetteGroup.map(b => b.bodyData.argOfPeriapsis!).sort((a, b) => a - b);
-    const expectedSpacing = 360 / rosetteGroup.length;
-
-    for (let i = 0; i < angles.length; i++) {
-      const nextIndex = (i + 1) % angles.length;
-      let spacing = angles[nextIndex] - angles[i];
-      if (spacing < 0) spacing += 360;
-      if (Math.abs(spacing - expectedSpacing) > ROSETTE_TOLERANCE_DEG) {
-        return null;
-      }
-    }
-
-    return `Rosette (${rosetteGroup.length})`;
-  }
-
-  /**
-   * Mean anomaly (degrees, wrapped to [0, 360)) propagated from the recorded sample
-   * to `now`. Inputs are assumed present; callers guard for missing orbital elements.
-   */
-  meanAnomalyNow(
-    meanAnomalyDeg: number,
-    orbitalPeriodDays: number,
-    meanAnomalyTimestamp: string,
-    now: number = Date.now(),
-  ): number {
-    const timestampMs = new Date(meanAnomalyTimestamp).getTime();
-    const elapsedDays = (now - timestampMs) / MS_PER_DAY;
-    const orbitalCycles = elapsedDays / orbitalPeriodDays;
-    // JS `%` keeps the sign of the dividend, so a negative mean anomaly (or a `now`
-    // before the sample timestamp) would otherwise return a value in (-360, 0). Add a
-    // full turn and re-wrap so the result is genuinely in [0, 360) as documented.
-    return (((meanAnomalyDeg + orbitalCycles * 360) % 360) + 360) % 360;
-  }
-
-  /** Walks to the root of the system tree containing `body`. */
-  private systemRoot(body: SystemBody): SystemBody {
-    let node = body;
-    while (node.parent) { node = node.parent; }
-    return node;
-  }
-
-  /** Every body in the system (root + all descendants), flattened. */
-  private flattenSystem(root: SystemBody): SystemBody[] {
-    const out: SystemBody[] = [root];
-    for (const child of root.subBodies) { out.push(...this.flattenSystem(child)); }
-    return out;
-  }
-
-  /**
-   * The most recent mean-anomaly observation timestamp recorded anywhere in `body`'s
-   * system — the shared epoch every body's displayed mean/true anomaly is calculated
-   * at, so bodies observed at different times still read as one consistent snapshot.
-   * Null when no body in the system carries a mean-anomaly timestamp.
-   */
-  systemAnomalyEpoch(body: SystemBody): Date | null {
-    let latestMs: number | null = null;
-    for (const b of this.flattenSystem(this.systemRoot(body))) {
-      const ts = b.bodyData.timestamps?.meanAnomaly;
-      const ms = ts ? Date.parse(ts) : NaN;
-      if (Number.isFinite(ms) && (latestMs === null || ms > latestMs)) { latestMs = ms; }
-    }
-    return latestMs === null ? null : new Date(latestMs);
-  }
-
-  /**
-   * Converts a mean anomaly (degrees) to a true anomaly (degrees, wrapped to [0, 360))
-   * for the given eccentricity by solving Kepler's equation M = E - e·sin E with
-   * Newton–Raphson, then mapping the eccentric anomaly E to the true anomaly ν. This
-   * gives the body's actual angular position along its orbit, measured from periapsis.
-   */
-  meanToTrueAnomaly(meanAnomalyDeg: number, eccentricity: number): number {
-    // Guard against non-finite inputs (e.g. a NaN eccentricity) so a single bad field
-    // can't poison the solver into returning NaN and silently freezing the live marker.
-    const e = Number.isFinite(eccentricity) ? Math.min(Math.max(eccentricity, 0), 0.999) : 0;
-    const meanDeg = Number.isFinite(meanAnomalyDeg) ? meanAnomalyDeg : 0;
-    const M = ((meanDeg % 360) + 360) % 360 * (Math.PI / 180);
-
-    // Newton–Raphson on f(E) = E - e·sin E - M. A handful of iterations converges to
-    // machine precision for all bound (e < 1) orbits.
-    let E = e < 0.8 ? M : Math.PI;
-    for (let i = 0; i < 12; i++) {
-      const delta = (E - e * Math.sin(E) - M) / (1 - e * Math.cos(E));
-      E -= delta;
-      if (Math.abs(delta) < 1e-10) break;
-    }
-
-    const trueAnomaly = 2 * Math.atan2(
-      Math.sqrt(1 + e) * Math.sin(E / 2),
-      Math.sqrt(1 - e) * Math.cos(E / 2),
-    );
-    return ((trueAnomaly * (180 / Math.PI)) % 360 + 360) % 360;
-  }
-
-  /** Degrees the body must still travel to reach its next apoapsis (180°) or periapsis (360°/0°). */
-  degreesToEvent(currentMeanAnomaly: number, type: 'apo' | 'peri'): number {
-    if (type === 'apo') {
-      let degrees = (180 - currentMeanAnomaly) % 360;
-      if (degrees < 0) { degrees += 360; }
-      return degrees;
-    }
-    return (360 - currentMeanAnomaly) % 360;
-  }
-
-  /**
-   * Next apoapsis/periapsis passage for a body, or null when the orbital elements are
-   * missing or the orbit is circular (no distinguishable apsis). Pure and time-injectable.
-   */
-  nextOrbitalEvent(bd: CanonnBiostatsBody, type: 'apo' | 'peri', now: number = Date.now()): OrbitalEvent | null {
-    // Use a null/undefined check for meanAnomaly (not a falsy check): a body that was
-    // exactly at periapsis at sample time has a legitimate meanAnomaly of 0, which a
-    // falsy check would wrongly treat as missing data and suppress the event row.
-    // `!bd.orbitalEccentricity` already excludes a circular orbit (e = 0, the "no
-    // distinguishable apsis" case) along with missing/NaN values.
-    if (bd.meanAnomaly == null || !bd.orbitalPeriod || !bd.timestamps?.meanAnomaly ||
-      !bd.orbitalEccentricity) {
-      return null;
-    }
-    const currentMeanAnomaly = this.meanAnomalyNow(bd.meanAnomaly, bd.orbitalPeriod, bd.timestamps.meanAnomaly, now);
-    const days = (this.degreesToEvent(currentMeanAnomaly, type) / 360) * bd.orbitalPeriod;
-    // A present-but-unparseable meanAnomaly timestamp makes `currentMeanAnomaly`
-    // (and thus `days`) NaN; bail out rather than emit an Invalid Date that the
-    // template's date pipe would throw on (NG02100/NG02311).
-    if (!Number.isFinite(days)) {
-      return null;
-    }
-    return { date: new Date(now + days * MS_PER_DAY), days };
-  }
-
-  /**
-   * Orbital radial range [periapsis, apoapsis] in AU, or null for an orbit that isn't a
-   * bound, recurring ellipse: a missing/non-positive semi-major axis, or an eccentricity
-   * ≥ 1 (parabolic/hyperbolic escape trajectory, which has no apoapsis and never returns,
-   * so it can't be a recurring collision partner). The `> 0` check is deliberate — a
-   * falsy/`!` test would let a negative semi-major axis (the convention for hyperbolic
-   * orbits, or corrupt external data) through and produce an inverted range.
-   */
-  private orbitalRadialRange(bd: CanonnBiostatsBody): { peri: number; apo: number } | null {
-    if (!(bd.semiMajorAxis! > 0)) { return null; }
-    const e = bd.orbitalEccentricity ?? 0;
-    if (!(e >= 0) || e >= 1) { return null; }
-    return { peri: bd.semiMajorAxis! * (1 - e), apo: bd.semiMajorAxis! * (1 + e) };
-  }
-
-  /**
-   * Parent-centric position (km) of a body at the given mean anomaly, from its Keplerian
-   * elements. Solves Kepler's equation for the eccentric anomaly, places the body in its
-   * orbital plane, then rotates by argument-of-periapsis, inclination and ascending node
-   * into the shared 3D frame. This is the geometry both the orbit-proximity test and the
-   * time-stepped collision search build on.
-   */
-  private orbitalStateVector(bd: CanonnBiostatsBody, meanAnomalyDeg: number): Vec3 {
-    const a = bd.semiMajorAxis! * KM_PER_AU;
-    const e = Math.min(Math.max(bd.orbitalEccentricity ?? 0, 0), 0.999);
-    const M = (((meanAnomalyDeg % 360) + 360) % 360) * DEG_TO_RAD;
-
-    let E = e < 0.8 ? M : Math.PI;
-    for (let i = 0; i < 12; i++) {
-      const delta = (E - e * Math.sin(E) - M) / (1 - e * Math.cos(E));
-      E -= delta;
-      if (Math.abs(delta) < 1e-12) { break; }
-    }
-
-    // Position in the orbital plane (periapsis along +x), then standard 3-1-3 rotation.
-    // The ED/Spansh data uses the opposite sign convention from the standard astronomical
-    // frame, so both angles must be negated before the rotation is applied.
-    const xo = a * (Math.cos(E) - e);
-    const yo = a * Math.sqrt(1 - e * e) * Math.sin(E);
-    const node = -(bd.ascendingNode ?? 0) * DEG_TO_RAD;
-    const argp = -(bd.argOfPeriapsis ?? 0) * DEG_TO_RAD;
-    const incl = (bd.orbitalInclination ?? 0) * DEG_TO_RAD;
-    const cO = Math.cos(node), sO = Math.sin(node);
-    const cw = Math.cos(argp), sw = Math.sin(argp);
-    const ci = Math.cos(incl), si = Math.sin(incl);
-    return {
-      x: xo * (cO * cw - sO * sw * ci) - yo * (cO * sw + sO * cw * ci),
-      y: xo * (sO * cw + cO * sw * ci) - yo * (sO * sw - cO * cw * ci),
-      z: xo * (sw * si) + yo * (cw * si),
-    };
-  }
-
-  /**
-   * Minimum distance (km) between the two orbit *curves*, independent of where each body
-   * currently is. When this exceeds the sum of the bodies' radii the orbits never touch in
-   * 3D — even if their radial ranges overlap, a relative tilt can hold the paths permanently
-   * apart — so the pair is not a collision candidate at all.
-   *
-   * Two near-circular orbits a few km apart approach within a sliver around their mutual
-   * node, far narrower than a uniform sweep can resolve, so the closest *coarse* sample alone
-   * reports a wildly inflated minimum. We defend against that two ways that don't depend on
-   * grid fineness: we keep the K closest coarse cells (multiple basins), and we additionally
-   * seed the refinement on the orbits' mutual line of nodes — where a relative tilt brings
-   * near-coplanar orbits closest. Every seed is then zoomed to sub-km precision, so the coarse
-   * grid only needs to land each true minimum in *some* seed's catchment, not resolve it. That
-   * lets the grid stay a fixed, bounded size ({@link ORBIT_COARSE_SAMPLES}) instead of an
-   * adaptive step that exploded the O(N²) double-loop for distant orbits.
-   */
-  private minOrbitDistanceKm(a: CanonnBiostatsBody, b: CanonnBiostatsBody): number {
-    // Fixed coarse grid: its job is to seed the refinement into the right basin, not to
-    // resolve the contact arc, so its cost is bounded regardless of orbital scale. The
-    // refinement window below is the coarse spacing, so each ±window covers the gap to the
-    // neighbouring sample and the zoom can reach any minimum between two coarse cells.
-    const stepDeg = 360 / ORBIT_COARSE_SAMPLES;
-
-    // Positions are computed once per sampled angle (per axis) and reused across the cross
-    // product, so an N×N grid costs 2N state-vector evaluations, not N².
-    const coarse: number[] = [];
-    for (let deg = 0; deg < 360; deg += stepDeg) { coarse.push(deg); }
-    const posA = coarse.map(d => this.orbitalStateVector(a, d));
-    const posB = coarse.map(d => this.orbitalStateVector(b, d));
-
-    // Keep the K closest coarse cells, not just the single closest. The closest approach of
-    // two near-coincident orbits lies in a narrow valley; the single best coarse sample can
-    // sit on a shallower stretch of that valley while the true minimum hides in a different
-    // cell, so we refine from several basins and take the overall minimum.
-    const topK: { d2: number; i: number; j: number }[] = [];
-    for (let i = 0; i < posA.length; i++) {
-      for (let j = 0; j < posB.length; j++) {
-        const dx = posA[i].x - posB[j].x, dy = posA[i].y - posB[j].y, dz = posA[i].z - posB[j].z;
-        const d2 = dx * dx + dy * dy + dz * dz;
-        if (topK.length < ORBIT_REFINE_TOPK) {
-          topK.push({ d2, i, j });
-          topK.sort((p, q) => q.d2 - p.d2); // worst first
-        } else if (d2 < topK[0].d2) {
-          topK[0] = { d2, i, j };
-          topK.sort((p, q) => q.d2 - p.d2);
-        }
-      }
-    }
-
-    // Refinement seeds: the K closest grid cells, plus the two ends of the mutual line of
-    // nodes (the line where the orbital planes intersect). For near-coplanar nested orbits the
-    // true minimum sits on that line — the grid can step over it, but the node seed cannot.
-    const seeds: [number, number][] = topK.map(cell => [coarse[cell.i], coarse[cell.j]]);
-    for (const seed of this.lineOfNodeSeeds(posA, posB, coarse)) { seeds.push(seed); }
-
-    let best = Infinity;
-    for (const [startA, startB] of seeds) {
-      best = Math.min(best, this.refineOrbitMinimum(a, b, startA, startB, stepDeg));
-    }
-    return best;
-  }
-
-  /**
-   * Mean-anomaly seeds at the two ends of the orbits' mutual line of nodes (the intersection
-   * of the two orbital planes), one (mean-anomaly-A, mean-anomaly-B) pair per end. For
-   * near-coplanar nested orbits the closest 3D approach lies on this line, where the relative
-   * tilt's vertical separation vanishes — a valley a uniform grid can step over. Returns []
-   * for effectively coplanar orbits: their planes are parallel, there is no distinct node
-   * line, and without a vertical valley the coarse grid already resolves the minimum.
-   */
-  private lineOfNodeSeeds(posA: Vec3[], posB: Vec3[], coarse: number[]): [number, number][] {
-    // Two position vectors from the focus span each orbital plane; their cross product is the
-    // plane normal. Use samples ~90° apart so they are well separated.
-    const quarter = Math.floor(coarse.length / 4);
-    const normalA = this.cross(posA[0], posA[quarter]);
-    const normalB = this.cross(posB[0], posB[quarter]);
-    const nodeLine = this.cross(normalA, normalB);
-    if (this.norm(nodeLine) <= 1e-6 * this.norm(normalA) * this.norm(normalB)) { return []; }
-
-    const seeds: [number, number][] = [];
-    for (const sign of [1, -1]) {
-      const dir: Vec3 = { x: sign * nodeLine.x, y: sign * nodeLine.y, z: sign * nodeLine.z };
-      seeds.push([this.mostAligned(posA, coarse, dir), this.mostAligned(posB, coarse, dir)]);
-    }
-    return seeds;
-  }
-
-  /** Mean anomaly of the sampled position whose direction (from the focus) best aligns with `dir`. */
-  private mostAligned(pos: Vec3[], coarse: number[], dir: Vec3): number {
-    let bestCos = -Infinity, bestIndex = 0;
-    for (let i = 0; i < pos.length; i++) {
-      const cos = this.dot(pos[i], dir) / this.norm(pos[i]);
-      if (cos > bestCos) { bestCos = cos; bestIndex = i; }
-    }
-    return coarse[bestIndex];
-  }
-
-  private cross(p: Vec3, q: Vec3): Vec3 {
-    return { x: p.y * q.z - p.z * q.y, y: p.z * q.x - p.x * q.z, z: p.x * q.y - p.y * q.x };
-  }
-  private dot(p: Vec3, q: Vec3): number { return p.x * q.x + p.y * q.y + p.z * q.z; }
-  private norm(p: Vec3): number { return Math.sqrt(this.dot(p, p)); }
-
-  /**
-   * Refines the orbit-to-orbit minimum distance starting from a coarse (mean-anomaly-A,
-   * mean-anomaly-B) cell, by repeatedly sampling a shrinking window around the running best.
-   */
-  private refineOrbitMinimum(a: CanonnBiostatsBody, b: CanonnBiostatsBody, startA: number, startB: number, stepDeg: number): number {
-    let aDeg = startA, bDeg = startB, best = Infinity;
-    let half = stepDeg;
-    for (let iter = 0; iter < ORBIT_REFINE_ITERATIONS; iter++) {
-      const degsA: number[] = [], degsB: number[] = [];
-      for (let k = -ORBIT_REFINE_GRID; k <= ORBIT_REFINE_GRID; k++) {
-        degsA.push(aDeg + (k / ORBIT_REFINE_GRID) * half);
-        degsB.push(bDeg + (k / ORBIT_REFINE_GRID) * half);
-      }
-      const posA = degsA.map(d => this.orbitalStateVector(a, d));
-      const posB = degsB.map(d => this.orbitalStateVector(b, d));
-      let cA = aDeg, cB = bDeg;
-      for (let i = 0; i < posA.length; i++) {
-        for (let j = 0; j < posB.length; j++) {
-          const dx = posA[i].x - posB[j].x, dy = posA[i].y - posB[j].y, dz = posA[i].z - posB[j].z;
-          const d = Math.sqrt(dx * dx + dy * dy + dz * dz);
-          if (d < best) { best = d; cA = degsA[i]; cB = degsB[j]; }
-        }
-      }
-      aDeg = cA; bDeg = cB;
-      half /= ORBIT_REFINE_GRID; // next window spans the previous grid spacing
-    }
-    return best;
-  }
-
-  /**
-   * Builds a closure giving the centre-to-centre distance (km) between two bodies at any
-   * epoch-ms time, by propagating each body's mean anomaly from its recorded sample and
-   * evaluating its 3D Keplerian position. Returns null when either body lacks the phase
-   * data (mean anomaly + timestamp) needed to place it in time. Shared by the collision
-   * search and the distance-over-time diagram so both measure separation identically.
-   */
-  private separationFunction(a: CanonnBiostatsBody, b: CanonnBiostatsBody): ((tMs: number) => number) | null {
-    if (
-      a.meanAnomaly == null || !a.orbitalPeriod || !a.timestamps?.meanAnomaly ||
-      b.meanAnomaly == null || !b.orbitalPeriod || !b.timestamps?.meanAnomaly
-    ) { return null; }
-
-    const epochA = Date.parse(a.timestamps.meanAnomaly);
-    const epochB = Date.parse(b.timestamps.meanAnomaly);
-    return (tMs: number): number => {
-      const Ma = a.meanAnomaly! + ((tMs - epochA) / MS_PER_DAY / a.orbitalPeriod!) * 360;
-      const Mb = b.meanAnomaly! + ((tMs - epochB) / MS_PER_DAY / b.orbitalPeriod!) * 360;
-      const pa = this.orbitalStateVector(a, Ma);
-      const pb = this.orbitalStateVector(b, Mb);
-      const dx = pa.x - pb.x, dy = pa.y - pb.y, dz = pa.z - pb.z;
-      return Math.sqrt(dx * dx + dy * dy + dz * dz);
-    };
-  }
-
-  /**
-   * Centre-to-centre distance (km) between two bodies sampled at `samples` evenly-spaced
-   * instants over [startMs, endMs] inclusive — the data behind the collision dialog's
-   * distance-over-time (synodic) diagram. The curve dips to a minimum at each conjunction
-   * and is deepest when that conjunction falls on the orbits' mutual node (a collision).
-   * Returns [] when either body lacks the phase data needed to place it in time, or when
-   * the window/sample count is degenerate.
-   */
-  separationSeries(a: CanonnBiostatsBody, b: CanonnBiostatsBody, startMs: number, endMs: number, samples: number): SeparationSample[] {
-    const sep = this.separationFunction(a, b);
-    if (!sep || !(endMs > startMs) || samples < 2) { return []; }
-    const step = (endMs - startMs) / (samples - 1);
-    const out: SeparationSample[] = [];
-    for (let i = 0; i < samples; i++) {
-      const tMs = startMs + i * step;
-      out.push({ tMs, sepKm: sep(tMs) });
-    }
-    return out;
-  }
-
-  /**
-   * Finds the time of minimum separation between two bodies within
-   * [centerMs − halfWindowMs, centerMs + halfWindowMs] using iterative zoom: 40 samples
-   * per pass, window halved and recentred on the running best each pass, until the
-   * half-window is under 500 ms (sub-second precision). This resolves arbitrarily narrow
-   * contact windows regardless of orbital scale.
-   */
-  /**
-   * Returns up to `count` upcoming contact windows between two bodies.
-   *
-   * Algorithm:
-   *   1. Coarse scan over [now − ½syn, now + 1syn] with 2000 equal steps to locate the
-   *      closest-approach spike (a sharp local minimum in the distance-vs-time curve).
-   *   2. Iteratively refine: centre a ±½syn window on the running best and halve each pass
-   *      until the half-window is under 500 ms (sub-second precision).
-   *   3. Step backward in synodic-period increments until the reference time is before now.
-   *   4. March forward one synodic period at a time; zoom-refine each event to sub-second
-   *      precision and record it as a contact when separation ≤ contactKm.
-   *
-   * Not every close approach is a collision — many conjunctions miss the orbits' mutual node
-   * — so each synodic event is evaluated individually rather than assumed to collide.
-   */
-  private nextContacts(a: CanonnBiostatsBody, b: CanonnBiostatsBody, contactKm: number, synodicDays: number, now: number, count: number, horizonMs: number = Infinity): CollisionWindow[] {
-    const sep = this.separationFunction(a, b);
-    if (!sep || !Number.isFinite(synodicDays)) { return []; }
-
-    const synodicMs = synodicDays * MS_PER_DAY;
-    // 2000 steps keeps the first-pass resolution (≈ synodic/2000 per step) comfortably
-    // below the narrowest expected contact window (~10 h for pairs with ~430-day synodics).
-    const STEPS = 2000;
-
-    /**
-     * Zoom to sub-second precision by repeatedly sampling STEPS points over a ±half window
-     * centred on the running best, halving each pass until half < 500 ms.
-     */
-    const zoomMin = (centerMs: number, halfMs: number): { t: number; sepKm: number } => {
-      let bestT = centerMs;
-      let bestS = Infinity;
-      let half = halfMs;
-      while (half > 500) {
-        const lo = bestT - half;
-        const hi = bestT + half;
-        const step = (hi - lo) / STEPS;
-        bestS = Infinity;
-        for (let t = lo; t <= hi; t += step) {
-          const s = sep(t);
-          if (s < bestS) { bestS = s; bestT = t; }
-        }
-        half /= 2;
-      }
-      return { t: bestT, sepKm: bestS };
-    };
-
-    // Step 1: coarse scan over [now − ½syn, now + 1syn] with STEPS equal steps.
-    const scanLo = now - synodicMs / 2;
-    const scanHi = now + synodicMs;
-    let initBestT = now;
-    let initBestS = Infinity;
-    const initStep = (scanHi - scanLo) / STEPS;
-    for (let t = scanLo; t <= scanHi; t += initStep) {
-      const s = sep(t);
-      if (s < initBestS) { initBestS = s; initBestT = t; }
-    }
-
-    // Step 2: refine the coarse best with ±½syn halving to sub-second precision.
-    const ref = zoomMin(initBestT, synodicMs / 2);
-
-    // Step 3: step backward in synodic-period increments until t < now.
-    let t0 = ref.t;
-    while (t0 >= now) { t0 -= synodicMs; }
-
-    // Step 4: march forward, zoom-refining each synodic event and collecting contacts.
-    const stepMs = (30 / 86400) * MS_PER_DAY; // 30-second probes for window bisection
-    const maxSpanMs = synodicMs / 2;           // bisection walk bound: half a synodic period
-    const results: CollisionWindow[] = [];
-
-    for (let k = 0; k < MAX_CONJUNCTIONS_SCANNED && results.length < count; k++) {
-      const candidateMs = t0 + k * synodicMs;
-      // Stop once conjunctions march past the caller's time horizon (later ones only recede
-      // further). Used by the simultaneity scan to bound work to the next N days.
-      if (candidateMs - now > horizonMs) { break; }
-      const min = zoomMin(candidateMs, synodicMs / 2);
-      // Allow min.t to be slightly before now: a contact whose minimum lands at now±ε
-      // (e.g. bodies aligned at the reference epoch) must not be discarded. We only skip
-      // conjunctions whose minimum is older than half a synodic period; contacts whose
-      // window has already ended are filtered below after computing endMs.
-      if (min.t < now - maxSpanMs || !Number.isFinite(min.sepKm)) { continue; }
-      if (min.sepKm > contactKm) { continue; }
-
-      // De-duplicate: skip if this minimum falls inside the last recorded contact window.
-      if (results.length > 0 && min.t <= results[results.length - 1].end.getTime()) { continue; }
-
-      // Root-find the contact window start (backward bisection from min.t).
-      let lo = min.t;
-      let hi = min.t - stepMs;
-      while (min.t - hi <= maxSpanMs && sep(hi) <= contactKm) { lo = hi; hi -= stepMs; }
-      for (let i = 0; i < 40; i++) {
-        const mid = (hi + lo) / 2;
-        if (sep(mid) > contactKm) { hi = mid; } else { lo = mid; }
-      }
-      const startMs = lo;
-      if (results.length > 0 && startMs <= results[results.length - 1].end.getTime()) { continue; }
-
-      // Root-find the contact window end (forward bisection from min.t).
-      let elo = min.t;
-      let ehi = min.t + stepMs;
-      while (ehi - min.t <= maxSpanMs && sep(ehi) <= contactKm) { elo = ehi; ehi += stepMs; }
-      for (let i = 0; i < 40; i++) {
-        const mid = (elo + ehi) / 2;
-        if (sep(mid) > contactKm) { ehi = mid; } else { elo = mid; }
-      }
-      const endMs = elo;
-
-      // Skip contacts whose window ended entirely before now (historical events; days < 0
-      // and the window is over). Contacts in progress (endMs > now, startMs ≤ now) are kept:
-      // days will be slightly negative, which CollisionWindow.days documents as intentional.
-      if (endMs < now) { continue; }
-
-      results.push({ start: new Date(startMs), end: new Date(endMs), days: (startMs - now) / MS_PER_DAY, minSeparationKm: min.sepKm, partnerName: b.name, combinedRadiiKm: contactKm });
-    }
-    return results;
-  }
-
-  /**
-   * Flags a body as a collision candidate when a sibling under the same parent is on a
-   * crossing orbit whose path comes within the sum of the two bodies' radii. The reported
-   * next-collision date is the next time the bodies are actually that close in 3D — not
-   * merely at the same longitude, since most conjunctions pass clear of the orbits' mutual
-   * intersection. Same-period co-orbital bodies (Trojans/rosettes) are excluded: they never
-   * lap each other (infinite synodic period) and are handled by the Trojan/rosette detectors.
-   *
-   * Note: this deliberately goes beyond the Canonn reference spreadsheet's 2D method (radial
-   * apo/periapsis-band overlap + synodic period, ignoring inclination, ascending node, and
-   * phase). The two agree for near-coplanar pairs but diverge on inclined pairs, where the 3D
-   * model defers the collision to the next conjunction that actually falls on the mutual node.
-   */
-  /**
-   * Every sibling under the same parent whose orbit actually crosses this body's within the
-   * combined radius — each a direct collision partner — with the pair's synodic period and
-   * contact threshold. Most bodies have at most one, but three or four siblings can share
-   * crossing orbits. Returns [] when this body itself isn't on a bound, recurring orbit.
-   */
-  private collisionPartners(body: SystemBody): { partner: SystemBody; synodic: number; contactKm: number }[] {
-    const bd = body.bodyData;
-    const range = this.orbitalRadialRange(bd);
-    if (!body.parent || !bd.orbitalPeriod || !range) { return []; }
-
-    const partners: { partner: SystemBody; synodic: number; contactKm: number }[] = [];
-    for (const sibling of body.parent.subBodies) {
-      if (sibling === body) { continue; }
-      const sd = sibling.bodyData;
-      const sRange = this.orbitalRadialRange(sd);
-      if (!sd.orbitalPeriod || !sRange) { continue; }
-      // Equal periods never lap (synodic period → ∞): these are stable co-orbital pairs.
-      if (sd.orbitalPeriod === bd.orbitalPeriod) { continue; }
-
-      // Cheap radial pre-filter: if the orbits' distance bands can't approach within the
-      // bodies' combined radius, they can never touch — skip the costly 3D work below.
-      const contactKm = (bd.radius ?? 0) + (sd.radius ?? 0);
-      const radialGapAu = Math.max(range.peri, sRange.peri) - Math.min(range.apo, sRange.apo);
-      if (radialGapAu > contactKm / KM_PER_AU) { continue; }
-
-      // Exact 3D candidacy: do the orbit curves themselves come within contact distance? A
-      // relative tilt can keep radially-overlapping orbits permanently apart.
-      if (this.minOrbitDistanceKm(bd, sd) > contactKm) { continue; }
-
-      const synodic = 1 / Math.abs(1 / bd.orbitalPeriod - 1 / sd.orbitalPeriod);
-      partners.push({ partner: sibling, synodic, contactKm });
-    }
-    return partners;
-  }
-
-  /**
-   * Every contact window for `body` across all its crossing partners that opens within the next
-   * `horizonDays`, in chronological order — uncapped, unlike {@link detectCollisionStatus}'s
-   * 10-row {@link CollisionStatus.upcomingCollisions}. Each window names its own partner. Lets the
-   * dialog mark *all* collisions inside the distance diagram's window, not just the first ten, so
-   * no in-view collision dip is left without a marker. Empty when the body has no crossing
-   * partners or lacks the phase data to time them.
-   */
-  upcomingContactsWithin(body: SystemBody, horizonDays: number, now: number = Date.now()): CollisionWindow[] {
-    return this.contactWindowsWithin(this.collisionPartners(body), body.bodyData, horizonDays * MS_PER_DAY, now)
-      .sort((a, b) => a.start.getTime() - b.start.getTime());
-  }
-
-  /**
-   * Contact windows opening within `horizonMs` for each given partner, merged (unsorted). The
-   * forward march is bounded both by the horizon and by MAX_CONJUNCTIONS_SCANNED, so tightly
-   * spaced pairs stay cheap.
-   */
-  private contactWindowsWithin(
-    partners: { partner: SystemBody; synodic: number; contactKm: number }[],
-    bodyData: CanonnBiostatsBody,
-    horizonMs: number,
-    now: number,
-  ): CollisionWindow[] {
-    const windows: CollisionWindow[] = [];
-    for (const p of partners) {
-      windows.push(...this.nextContacts(bodyData, p.partner.bodyData, p.contactKm, p.synodic, now, MAX_CONJUNCTIONS_SCANNED, horizonMs)
-        .filter(w => w.start.getTime() <= now + horizonMs));
-    }
-    return windows;
-  }
-
-  /**
-   * Timed multi-body pile-ups for `body` over the next `horizonDays`: intervals in which it is
-   * simultaneously within contact of two or more siblings. Unlike the simultaneity derived from
-   * {@link detectCollisionStatus}'s capped upcoming-contacts list, this scans every direct
-   * partner's contacts across the whole horizon, so a cluster beyond the listed rows is still
-   * found. Empty when the body has fewer than two crossing partners.
-   */
-  simultaneousCollisionsWithin(body: SystemBody, horizonDays: number, now: number = Date.now()): SimultaneousCollision[] {
-    const partners = this.collisionPartners(body);
-    if (partners.length < 2) { return []; }
-    return this.groupSimultaneous(this.contactWindowsWithin(partners, body.bodyData, horizonDays * MS_PER_DAY, now), now);
-  }
-
-  /**
-   * Groups contact windows that overlap in time; a group spanning two or more distinct partners
-   * is a simultaneous multi-body collision. Mirrors the time-overlap clustering the dialog does
-   * on its visible rows, but over the full horizon set.
-   */
-  private groupSimultaneous(windows: CollisionWindow[], now: number): SimultaneousCollision[] {
-    const sorted = [...windows].sort((a, b) => a.start.getTime() - b.start.getTime());
-    const out: SimultaneousCollision[] = [];
-    let group: CollisionWindow[] = [];
-    let maxEnd = -Infinity;
-
-    const flush = (): void => {
-      const partnerNames = new Set(group.map(w => w.partnerName).filter((n): n is string => !!n));
-      if (partnerNames.size >= 2) {
-        out.push({
-          partnerNames: [...partnerNames].sort(),
-          start: new Date(Math.min(...group.map(w => w.start.getTime()))),
-          end: new Date(Math.max(...group.map(w => w.end.getTime()))),
-          days: Math.min(...group.map(w => w.days)),
-        });
-      }
-      group = [];
-    };
-
-    for (const w of sorted) {
-      if (group.length === 0 || w.start.getTime() <= maxEnd) {
-        group.push(w);
-        maxEnd = Math.max(maxEnd, w.end.getTime());
-      } else {
-        flush();
-        group = [w];
-        maxEnd = w.end.getTime();
-      }
-    }
-    flush();
-    return out;
-  }
-
-  detectCollisionStatus(body: SystemBody, now: number = Date.now()): CollisionStatus {
-    const none: CollisionStatus = {
-      isCandidate: false, partnerName: null, synodicPeriodDays: null, nextCollision: null, upcomingCollisions: [], combinedRadiiKm: null, simultaneousPartners: [],
-    };
-    const bd = body.bodyData;
-    const range = this.orbitalRadialRange(bd);
-    if (!body.parent || !bd.orbitalPeriod || !range) { return none; }
-
-    const partners = this.collisionPartners(body);
-    if (partners.length === 0) { return none; }
-
-    // Compute each partner's upcoming contact windows, then merge them into one chronological
-    // list so the soonest collisions surface regardless of which sibling they involve. Each
-    // window already carries its partner's name and contact radius (stamped in nextContacts).
-    const merged: CollisionWindow[] = [];
-    for (const p of partners) {
-      merged.push(...this.nextContacts(bd, p.partner.bodyData, p.contactKm, p.synodic, now, MAX_UPCOMING_CONTACTS));
-    }
-    merged.sort((x, y) => x.start.getTime() - y.start.getTime());
-    const upcoming = merged.slice(0, MAX_UPCOMING_CONTACTS);
-
-    // The primary partner owns the soonest collision; fall back to the first geometric
-    // candidate when no pair has usable phase/timing data (so the badge still appears).
-    const primary = (upcoming[0]
-      ? partners.find(p => p.partner.bodyData.name === upcoming[0].partnerName)
-      : null) ?? partners[0];
-
-    // Identify additional siblings that are part of the same crossing-orbit group, making
-    // this a multi-body cluster. Grow the group transitively from every direct partner: if
-    // A crosses B and B crosses C, C is in the group even if A doesn’t directly cross C.
-    const groupMembers = new Set<string>([bd.name, ...partners.map(p => p.partner.bodyData.name)]);
-    let changed = true;
-    while (changed) {
-      changed = false;
-      for (const sibling of body.parent.subBodies) {
-        if (groupMembers.has(sibling.bodyData.name)) { continue; }
-        const sd = sibling.bodyData;
-        const sRange = this.orbitalRadialRange(sd);
-        if (!sd.orbitalPeriod || !sRange) { continue; }
-        for (const memberName of groupMembers) {
-          const member = memberName === bd.name
-            ? body
-            : body.parent!.subBodies.find(s => s.bodyData.name === memberName);
-          if (!member) { continue; }
-          const md = member.bodyData;
-          const mRange = this.orbitalRadialRange(md);
-          if (!mRange || !md.orbitalPeriod || md.orbitalPeriod === sd.orbitalPeriod) { continue; }
-          const contactKm = (md.radius ?? 0) + (sd.radius ?? 0);
-          const radialGapAu = Math.max(mRange.peri, sRange.peri) - Math.min(mRange.apo, sRange.apo);
-          if (radialGapAu > contactKm / KM_PER_AU) { continue; }
-          if (this.minOrbitDistanceKm(md, sd) > contactKm) { continue; }
-          groupMembers.add(sibling.bodyData.name);
-          changed = true;
-          break;
-        }
-      }
-    }
-    const simultaneousPartners = [...groupMembers].filter(n => n !== bd.name && n !== primary.partner.bodyData.name);
-
-    return {
-      isCandidate: true,
-      partnerName: primary.partner.bodyData.name,
-      synodicPeriodDays: primary.synodic,
-      nextCollision: upcoming[0] ?? null,
-      upcomingCollisions: upcoming,
-      combinedRadiiKm: primary.contactKm,
-      simultaneousPartners,
-    };
-  }
-}
+export * from './orbital-relations.core';
