+++ src/app/system-body/system-body.component.html
@@ -79,13 +79,18 @@
       <span class="badge badge-red" [matTooltip]="'Klemperer ' + rosetteStatus">{{
         rosetteStatus }}</span>
       }
-      @if (collisionStatus?.isCandidate) {
-      <span class="badge badge-red clickable" [matTooltip]="collisionStatus!.nextCollision
-              ? ('Collides with ' + collisionStatus!.partnerName
-                 + ' in ' + formatCollisionCountdown(collisionStatus!.nextCollision!.days)
-                 + ' — click for details')
-              : ('Crossing orbit with ' + collisionStatus!.partnerName + ' — click for details')"
-        (click)="showCollisionDialog()">{{ (collisionStatus!.nextCollision?.days ?? 0) < 0 ? 'Collision In Progress' : 'Collision' }}</span>
+      @if (collisionStatus(); as cs) {
+        @if (cs.isCandidate) {
+        <span class="badge badge-red clickable" [matTooltip]="cs.nextCollision
+                ? ('Collides with ' + cs.partnerName
+                   + ' in ' + formatCollisionCountdown(cs.nextCollision.days)
+                   + ' — click for details')
+                : ('Crossing orbit with ' + cs.partnerName + ' — click for details')"
+          (click)="showCollisionDialog()">{{ (cs.nextCollision?.days ?? 0) < 0 ? 'Collision In Progress' : 'Collision' }}</span>
+        }
+      } @else if (collisionPending()) {
+        <span class="badge badge-skeleton" aria-hidden="true"
+          matTooltip="Checking for crossing orbits…">Collision</span>
       }
       @if (getRocheExcess() !== null) {
       <span class="badge badge-red"
